+++ common/test/tests_vdev.c
@@ -2,7 +2,6 @@
 #include <gic_regs.h>
 #include <asm-arm_inline.h>
 #include <gic.h>
-#include "virqmap.h"
 #include <config/cfg_platform.h>
 #include <log/print.h>
