+++ hypervisor/hal/arm32ve/guest_hal.c
@@ -7,7 +7,7 @@
 #include <mm.h>
 #include <interrupt.h>
 #include <vdev.h>
-#include <virqmap.h>
+#include <virq.h>
 #include <guest.h>
 #include <guest_hal.h>
 #include <vmm.h>
