+++ hypervisor/hal/arm32ve/include/timer_hal.h
