+++ hypervisor/hal/arm32ve/libhw/gic.h
