+++ hypervisor/hal/arm32ve/libhw/lpae.c
