+++ hypervisor/hal/arm32ve/libhw/lpae.h
