+++ hypervisor/hal/arm32ve/libhw/slotpirq.h
