+++ hypervisor/hal/arm32ve/libhw/trap.c
