+++ hypervisor/hal/arm32ve/libhw/trap.h
