+++ hypervisor/hal/arm32ve/libhw/vector.S
