+++ hypervisor/hal/arm32ve/libhw/vgic.c
