+++ hypervisor/hal/arm32ve/vdev/vdev_cp.c
