+++ hypervisor/hal/arm32ve/vdev/vdev_sample.c
@@ -5,7 +5,7 @@
 #define DEBUG
 #include <log/print.h>
 #include <asm-arm_inline.h>
-#include <virqmap.h>
+#include <virq.h>
 
 #define SAMPLE_BASE_ADDR 0x3FFFF000
