+++ hypervisor/include/mm.h
