+++ hypervisor/include/timer.h
