+++ hypervisor/include/vmm.h
